pyproject.toml
[project]
name = "superbot"
version = "0.1.0"
description = "AI Assistant for Superteam Vietnam"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "bitsandbytes>=0.45.1",
    "fastapi>=0.115.7",
    "httpx>=0.28.1",
    "langchain>=0.3.15",
    "langchain-chroma>=0.2.0",
    "langchain-huggingface>=0.1.2",
    "llama-cpp-python>=0.3.6",
    "loguru>=0.7.3",
    "pre-commit>=4.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pymupdf>=1.25.2",
    "pypdf>=5.1.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "python-multipart>=0.0.20",
    "python-telegram-bot>=21.10",
    "rank-bm25>=0.2.2",
    "redis>=5.2.1",
    "ruff>=0.9.3",
    "sentence-transformers>=3.4.0",
    "sqlmodel>=0.0.22",
    "testcontainers>=4.9.1",
    "transformers>=4.48.1",
    "tweepy>=4.15.0",
    "uvicorn[standard]>=0.34.0",
    "weaviate-client>=4.10.4",
]


[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80
quote-style = "single"
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # flake8-bugbear
    "B",
]




# [build-system]
# requires = ["setuptools>=65.5.1"]  # Only setuptools
# build-backend = "setuptools.build_meta"
